# refactor int_gauss_hermite_term