This is a refactor of a single function, `int_gauss_hermite_term`.  My goal here was primarily to make it more concise and maintainable. 

There is also a speedup for the $n \neq 0$ cases, but that comes almost entirely from using the `sqrt` function instead of `^0.5` in the gaussian pdf calculation.  It's a little slower in the $n=0$ case, but faster otherwise, which should be a net speedup.

<details><summary>timing details</summary>

Time to compute CDF and PDF for all supported $n$.

```
n = 0
new implementation:        0.000037 seconds (16 allocations: 24.094 KiB)
original implementation:   0.000030 seconds (7 allocations: 9.656 KiB)

n = 1
new implementation:        0.000015 seconds (19 allocations: 28.906 KiB)
original implementation:   0.000028 seconds (32 allocations: 29.562 KiB)

n = 2
new implementation:        0.000015 seconds (19 allocations: 28.906 KiB)
original implementation:   0.000034 seconds (28 allocations: 24.969 KiB)

n = 3
new implementation:        0.000016 seconds (19 allocations: 28.906 KiB)
original implementation:   0.000029 seconds (29 allocations: 25.031 KiB)

n = 4
new implementation:        0.000015 seconds (19 allocations: 28.906 KiB)
original implementation:   0.000026 seconds (32 allocations: 29.562 KiB)

n = 5
new implementation:        0.000014 seconds (19 allocations: 28.906 KiB)
original implementation:   0.000035 seconds (32 allocations: 29.562 KiB)

n = 6
new implementation:        0.000024 seconds (19 allocations: 28.906 KiB)
original implementation:   0.000043 seconds (32 allocations: 29.562 KiB)

n = 7
new implementation:        0.000015 seconds (19 allocations: 28.906 KiB)
original implementation:   0.000046 seconds (32 allocations: 29.562 KiB)

n = 8
new implementation:        0.000016 seconds (19 allocations: 28.906 KiB)
original implementation:   0.000056 seconds (32 allocations: 29.562 KiB)

n = 9
new implementation:        0.000016 seconds (19 allocations: 28.906 KiB)
original implementation:   0.000059 seconds (32 allocations: 29.562 KiB)

n = 10
new implementation:        0.000015 seconds (19 allocations: 28.906 KiB)
original implementation:   0.000068 seconds (32 allocations: 29.562 KiB)
```

</details>