@ajwheeler 2025-03-27
new - old:
![image](https://github.com/user-attachments/assets/e66043d0-87ec-4d93-a852-70061de541a9)